Add tests for array coercion and split existing ones
