+++ frontend/test/resolution/CMakeLists.txt
@@ -54,6 +54,8 @@ comp_unit_test(testInteractive)
 comp_unit_test(testInterfaces)
 comp_unit_test(testIterators)
 comp_unit_test(testIteratorsToArrays)
+comp_unit_test(testIteratorsToArraysCoerce)
+comp_unit_test(testIteratorsToArraysWithPromotion)
 comp_unit_test(testLet)
 comp_unit_test(testLibrary)
 comp_unit_test(testLocation)
